pyproject.toml
[project]
name = "robotdreams-home-work"
version = "0.1.0"
description = "home work for robotdreams"
requires-python = ">=3.12"
dependencies = [
    "fastavro>=1.12.1",
    "flask>=3.1.2",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
]

[tool.uv]
dev-dependencies = [
    "pre-commit",
    "black",
    "isort",
    "ruff",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88
